First commit to new branch. The only stuff that is known to compile are:
Original commit message from CVS:
First commit to new branch. The only stuff that is known to compile are:
- the ext/mad dir
- the gst/mpegaudioparse dir
- the sys/oss dir
Why these? - because I hacked them up to do seeking in Mp3's with the player - The changes are not supposed to stay there. At least the osssink hack is ugly, but I wanted easy output :)